flink2.0/requirements.txt
kafka-python
redis
numpy
sentence-transformers
python-dotenv
